Add models for describing output facing package data
repo_management/models.py:
Add model `PackageFiles` to track the list of files for a given package.
Add model `RepoDbMemberType` to track the type of a repository database
member (distinguished by the IntEnum defaults.RepoDbMemberType).
Add model `RepoDbMemberData` to track the data (represented by an
io.StringIO instance) of a repository database member.
Add model `OutputPackage` to track the attributes required for a single
package in the context of a representation within its pkgbase.
Add model `OutputPackageBase` to track the attributes required for a
pkgbase and its list of packages.