Add iterator over repository database
repo_management/operations.py:
Add the public method `db_file_as_models()`, which yields all members of
a repository database as a Tuple of the pkgbase name and the pkgbase
data (represented as an instance of models.OutputPackageBase).

tests/test_operations.py:
Add a test for `db_file_as_models()`.