Add a set of error classes
repo_management/errors.py:
Add the Exception class `RepoManagementError` serving as a base, from
which `RepoManagementFileError` is derived, which serves as a general
Error class to signal problems with file interaction.
Add the Exception class `RepoManagementFileNotFoundError`, which is
derived from `RepoManagementFileError`.
Add `RepoManagementValidationError` which derives from
`RepoManagementError` and is raised on validation issues of file
data.