tox: fix poetry usage
Currently we are using tox, a virtual environment orchestrator, to run
an external poetry command, which has its own virtual environment.
This makes tests slow, as we don't get tox's optimizations, and we are
provisioning two virtual environments, one in tox and one in poetry.

This patch remove the external poetry usages and makes use of tox as it
was designed. It uses the tox-poetry-dev-dependencies[1] plugin, which
does not need to be installed on the host, to install the dev
dependencies, as for some reason poetry does not export these via an
extra, as other backends normally do.
The tests are now much faster, as we skip provisioning the poetry
environment entirely, and we make use of tox's caching and other
optimizations.

[1] https://github.com/sinoroc/tox-poetry-dev-dependencies

Signed-off-by: Filipe Laíns <lains@riseup.net>