Added a 'getting started' guide. It most certenly is wrong either as a whole concept, or parts of it. In which case, feel free to scrap - change - or point it out :)
Revert "Added a 'getting started' guide. It most certenly is wrong either as a whole concept, or parts of it. In which case, feel free to scrap - change - or point it out :)"

This reverts commit 53cc3b10505abb18c95d33ef3431d440bd6e7530.

Added a getting started guide. I used markdown to get the code blocks working, as .rst didn't seem to work with .. code:: blocks. It most certenly is wrong either as a whole concept, or parts of it. In which case, feel free to scrap - change - or point it out :)

Changed getting started to something more slim, as per discussion with @dvzrv

Merged suggestions from @artafinde

Added a 'getting started' guide. It most certenly is wrong either as a whole concept, or parts of it. In which case, feel free to scrap - change - or point it out :)

Revert "Added a 'getting started' guide. It most certenly is wrong either as a whole concept, or parts of it. In which case, feel free to scrap - change - or point it out :)"

This reverts commit 53cc3b10505abb18c95d33ef3431d440bd6e7530.

Added a getting started guide. I used markdown to get the code blocks working, as .rst didn't seem to work with .. code:: blocks. It most certenly is wrong either as a whole concept, or parts of it. In which case, feel free to scrap - change - or point it out :)

Changed getting started to something more slim, as per discussion with @dvzrv

Merged suggestions from @artafinde

Updated to use poetry instead of installed db2json for testing.