Use pyalpm's vercmp if it is available
repod/common/models.py:
Attempt to import pyalpm's `vercmp()` and use it in `PkgVer.vercmp()`
and `Version.vercmp()` instead of the builtin implementation if it is
available.

tests/common/test_models.py:
Parametrize tests for `PkgVer.vercmp()` and `Version.vercmp()` further
so that they can run against pyalpm's `vercmp()` and the builtin
implementation.