Use poetry in tox again
tox.ini:
Call poetry in tox.ini again, as this way we can ensure to install
all packages based on the lock file.
Additionally, this allows to properly and selectively install extras,
required e.g. by the coverage target (as to not install pyalpm for all
targets).

.gitlab-ci.yml:
Install python-poetry to test environment.

pyproject.toml:
Add setuptools to build requirements because of bug in pip:
https://github.com/pypa/pip/issues/6100