Return List instead of AsyncIterator with SyncDatabase.outputpackagebases()
repod/repo/package/syncdb.py:
Return a `List` instead of an `AsyncIterator` over the names and
OutputPackageBase tuples, as that is easier to use with `asyncio.run()`.

repod/operations.py:
Adapt calls to `SyncDatabase.outputpackagebases()` to expect a List and
not an AsyncIterator.

tests/repo/package/test_syncdb.py:
Adapt tests for `SyncDatabase.outputpackagebases()` to reflect the
changed return type.