Add CI logic for testing the latest dependencies
This way we can catch breaking upstream changes early and adapt the code
sooner rather than later.

It works by checking if the CI variable TEST_LATEST is set to "true" all
dependencies is bumped to the latest version, before tests are run. The
plan is to test the latest dependencies on a schedule.

Fix #67