Exchange Union for type union operator and Optional for | None
This was introduced in 3.10 and allows to use the set union operator (the pipe)
to be used instead of Union, e.g. Union[foo, bar] can be expressed as foo | bar.

Optional can now be expressed using this, e.g. Optional[int] as int | None,
which is shorter, doesn't need an import and is less confusing when a required
argument is not typed as "optional".

https://docs.python.org/3/whatsnew/3.10.html#pep-604-new-type-union-operator