Do not look up build requirements in integration tests
tests/cli/test_cli.py:
Change `test_import_into_default_repo()` to run the package import
without checking for build requirements (as we would need access to the
archive to run this test).
Change the test to fail on import failing and to provide debug
information.
Lower amount of imported packages to 5 to not blow up the test cache too
much.